Refactored settings:ui and :common:settings into dedicated settings feature module
