--- common/settings/src/commonMain/kotlin/app.campfire.common.settings/SettingsComponent.kt
@@ -1,9 +1,0 @@
-package app.campfire.common.settings
-
-import app.campfire.core.di.AppScope
-import com.r0adkll.kimchi.annotations.ContributesTo
-
-expect interface PreferencesPlatformComponent
-
-@ContributesTo(AppScope::class)
-interface SettingsComponent : PreferencesPlatformComponent
