--- common/settings/src/jvmMain/kotlin/app/campfire/common/settings/PreferencesPlatformComponent.kt
@@ -1,15 +1,0 @@
-package app.campfire.common.settings
-
-import app.campfire.core.di.AppScope
-import app.campfire.core.di.SingleIn
-import com.russhwolf.settings.ObservableSettings
-import com.russhwolf.settings.PreferencesSettings
-import java.util.prefs.Preferences
-import me.tatarka.inject.annotations.Provides
-
-actual interface PreferencesPlatformComponent {
-
-  @SingleIn(AppScope::class)
-  @Provides
-  fun provideSettings(delegate: Preferences): ObservableSettings = PreferencesSettings(delegate)
-}
