+++ data/account/impl/build.gradle.kts
@@ -12,7 +12,7 @@ kotlin {
     commonMain {
       dependencies {
         api(projects.data.account.api)
-        api(projects.common.settings)
+        api(projects.features.settings.api)
         api(projects.data.db)
         api(projects.data.network.api)
         api(projects.data.mapping)
