+++ features/search/impl/build.gradle.kts
@@ -15,7 +15,7 @@ kotlin {
         api(projects.features.search.api)
         implementation(projects.features.user.api)
 
-        implementation(projects.common.settings)
+        implementation(projects.features.settings.api)
         implementation(projects.core)
         implementation(projects.data.db)
         implementation(projects.data.network.api)
