+++ features/sessions/impl/build.gradle.kts
@@ -18,7 +18,7 @@ kotlin {
       dependencies {
         api(projects.features.sessions.api)
 
-        implementation(projects.common.settings)
+        implementation(projects.features.settings.api)
         implementation(projects.core)
         implementation(projects.data.db)
         implementation(projects.data.network.api)
