+++ features/settings/impl/src/androidMain/kotlin/app/campfire/settings/PreferencesPlatformComponent.kt
@@ -1,0 +1,29 @@
+package app.campfire.settings
+
+import android.app.Application
+import android.content.SharedPreferences
+import androidx.preference.PreferenceManager
+import app.campfire.core.di.AppScope
+import app.campfire.core.di.SingleIn
+import com.russhwolf.settings.ObservableSettings
+import com.russhwolf.settings.SharedPreferencesSettings
+import me.tatarka.inject.annotations.Provides
+
+actual interface PreferencesPlatformComponent {
+
+  @SingleIn(AppScope::class)
+  @Provides
+  fun provideSettings(delegate: AppSharedPreferences): ObservableSettings {
+    return SharedPreferencesSettings(delegate)
+  }
+
+  @SingleIn(AppScope::class)
+  @Provides
+  fun provideAppPreferences(
+    context: Application,
+  ): AppSharedPreferences =
+    PreferenceManager.getDefaultSharedPreferences(context)
+}
+
+
+typealias AppSharedPreferences = SharedPreferences
