+++ features/user/impl/build.gradle.kts
@@ -12,7 +12,7 @@ kotlin {
   sourceSets {
     commonMain {
       dependencies {
-        implementation(projects.common.settings)
+        implementation(projects.features.settings.api)
         implementation(projects.core)
         implementation(projects.data.db)
         implementation(projects.data.network.api)
