--- ui/settings/build.gradle.kts
@@ -1,29 +1,0 @@
-plugins {
-  id("app.campfire.ui")
-}
-
-kotlin {
-  sourceSets {
-    commonMain {
-      dependencies {
-        api(projects.common.compose)
-        api(projects.data.account.api)
-        api(projects.data.account.ui)
-
-        implementation(compose.components.resources)
-      }
-    }
-
-    jvmMain {
-      dependencies {
-        implementation(compose.preview)
-      }
-    }
-
-    androidMain {
-      dependencies {
-        implementation(compose.preview)
-      }
-    }
-  }
-}
